Move close contacts computation to custom calculator class
